gather/requirements.txt
iso8601==1.0.2
PyMySQL==1.0.2
requests==2.27.1
python-dotenv
